+++ src/pages/sunshine/[entity]/[id]/costs-and-tariffs.tsx
@@ -739,6 +739,8 @@ const CostsAndTariffs = (props: Props) => {
         SidebarContent={sidebarContent}
         MainContent={mainContent}
         download={query.download}
+        entity={entity}
+        id={id}
       />
     </>
   );