+++ src/pages/sunshine/[entity]/[id]/overview.tsx
@@ -702,6 +702,8 @@ const OverviewPage = (props: Props) => {
           SidebarContent={sidebarContent}
           MainContent={mainContent}
           download={query.download}
+          entity={entity}
+          id={id}
         />
       </SafeHydration>
     </>