+++ src/pages/sunshine/[entity]/[id]/power-stability.tsx
@@ -514,6 +514,8 @@ const PowerStability = (props: Props) => {
         SidebarContent={sidebarContent}
         MainContent={mainContent}
         download={query.download}
+        entity={entity}
+        id={id}
       />
     </>
   );